Separated out deployment files from JS API
